refactor(llma): address Carlos's review feedback
- Let middleware decide whether to run via matches() instead of
  marker keys in the registry, enabling stacking and version-specific
  matching (Express.js style)
- Extract event name classification into event_name.rs, following the
  same pattern as identity.rs in the otel folder
