+++ nodejs/src/ingestion/ai/otel/index.ts
@@ -8,17 +8,8 @@ import { vercelAi } from './middleware/vercel-ai'
 // Middleware registry — checked in order, first match wins.
 const MIDDLEWARES: OtelLibraryMiddleware[] = [pydanticAi, vercelAi]
 
-function detectMiddleware(event: PluginEvent): OtelLibraryMiddleware | undefined {
-    for (const mw of MIDDLEWARES) {
-        if (mw.markerKeys.some((key) => event.properties?.[key] !== undefined)) {
-            return mw
-        }
-    }
-    return undefined
-}
-
 export function convertOtelEvent(event: PluginEvent): void {
-    const middleware = detectMiddleware(event)
+    const middleware = MIDDLEWARES.find((mw) => mw.matches(event))
 
     if (middleware) {
         middleware.process(event, () => mapOtelAttributes(event))
