+++ nodejs/src/ingestion/ai/otel/middleware/pydantic-ai.ts
@@ -128,8 +128,9 @@ function process(event: PluginEvent, next: () => void): void {
     }
 }
 
+const MARKER_KEYS = ['pydantic_ai.all_messages', 'logfire.msg', 'logfire.json_schema', 'model_request_parameters']
+
 export const pydanticAi: OtelLibraryMiddleware = {
-    name: 'opentelemetry/pydantic-ai',
-    markerKeys: ['pydantic_ai.all_messages', 'logfire.msg', 'logfire.json_schema', 'model_request_parameters'],
+    matches: (event) => MARKER_KEYS.some((key) => event.properties?.[key] !== undefined),
     process,
 }
