+++ nodejs/src/ingestion/ai/otel/middleware/vercel-ai.ts
@@ -141,8 +141,9 @@ function process(event: PluginEvent, next: () => void): void {
     }
 }
 
+const MARKER_KEYS = ['ai.operationId', 'ai.telemetry.functionId']
+
 export const vercelAi: OtelLibraryMiddleware = {
-    name: 'opentelemetry/vercel-ai',
-    markerKeys: ['ai.operationId', 'ai.telemetry.functionId'],
+    matches: (event) => MARKER_KEYS.some((key) => event.properties?.[key] !== undefined),
     process,
 }
